backend/pyproject.toml
[project]
name = "baulv-backend"
version = "0.1.0"
description = "AI-gestützte Bau-Ausschreibungssoftware Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.9",
    "anthropic>=0.40.0",
    "pymupdf>=1.24.0",
    "pdf2image>=1.17.0",
    "openpyxl>=3.1.0",
    "pgvector>=0.3.0",
    "httpx>=0.27.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "sse-starlette>=2.0.0",
    "Pillow>=10.0.0",
    "stripe>=7.0.0",
    "bcrypt>=4.0.0",
    "email-validator>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "factory-boy>=3.3.0",
    "httpx>=0.27.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
